cambio de datos a long y se han agregado servicios rest administrador

+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/auxentities/UserCreationModel.java
@@ -6,7 +6,7 @@ public class UserCreationModel {
 	private String user;
 	private String password;
 	private String mail;
-	private int id_client;
+	private long id_client;
 	private int profile;
 	private String payment_type;
 	private String patente_vehiculo;
@@ -36,10 +36,10 @@ public class UserCreationModel {
 	public void setMail(String mail) {
 		this.mail = mail;
 	}
-	public int getId_client() {
+	public long getId_client() {
 		return id_client;
 	}
-	public void setId_client(int id_client) {
+	public void setId_client(long id_client) {
 		this.id_client = id_client;
 	}
 	public int getProfile() {
