+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/entities/Gps.java
@@ -19,10 +19,10 @@ public class Gps {
 	
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
-	private int id;
+	private long id;
 	
 	@Column(name = "id_user", unique = false, nullable = false)
-	private int id_user;
+	private long id_user;
 	
 	@Column(name = "latitude", unique = false, nullable = false)
 	private double latitude;
@@ -35,14 +35,14 @@ public class Gps {
 	@JsonIgnore
 	private User user_gps;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
-	public int getId_user() {
+	public long getId_user() {
 		return id_user;
 	}
-	public void setId_user(int id_user) {
+	public void setId_user(long id_user) {
 		this.id_user = id_user;
 	}
