+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/entities/Gps_last.java
@@ -20,10 +20,10 @@ private static final long serialVersionUID = 1L;
 	
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
-	private int id;
+	private long id;
 	
 	@Column(name = "id_user", unique = false, nullable = false)
-	private int id_user;
+	private long id_user;
 	
 	@Column(name = "latitude", unique = false, nullable = false)
 	private double latitude;
@@ -37,15 +37,15 @@ private static final long serialVersionUID = 1L;
 	private User user_gps_last;
 	
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
-	public int getId_user() {
+	public long getId_user() {
 		return id_user;
 	}
 
-	public void setId_user(int id_user) {
+	public void setId_user(long id_user) {
 		this.id_user = id_user;
 	}
