+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/entities/Lista.java
@@ -23,7 +23,7 @@ public class Lista {
 	
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
-	private int id;
+	private long id;
 	
 	@ManyToOne
 	@JoinColumn(name = "id_user", insertable = false, updatable = false)
@@ -31,7 +31,7 @@ public class Lista {
 	private User user_list;
 	
 	@Column(name = "id_user", unique = false, nullable = false)
-	private int id_user;
+	private long id_user;
 	
 	@Column(name = "price", unique = false, nullable = false)
 	private int price;
@@ -42,15 +42,15 @@ public class Lista {
 	@OneToMany(mappedBy = "lista")
 	private List<Detail> detail_list;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	
-	public int getId_user() {
+	public long getId_user() {
 		return id_user;
 	}
 
-	public void setId_user(int id_user) {
+	public void setId_user(long id_user) {
 		this.id_user = id_user;
 	}
