+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/ClientRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import cl.apicom.spring.backend.entities.Client;
 
-public interface ClientRepository extends PagingAndSortingRepository<Client, Integer>{
+public interface ClientRepository extends PagingAndSortingRepository<Client, Long>{
 	
 	@Query("SELECT c FROM Client c where id != 0")
 	List<Client> getAllExceptFirst();
