+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/ListRepository.java
@@ -3,6 +3,6 @@ package cl.apicom.spring.backend.repository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import cl.apicom.spring.backend.entities.Lista;
 
-public interface ListRepository extends PagingAndSortingRepository<Lista, Integer>{
+public interface ListRepository extends PagingAndSortingRepository<Lista, Long>{
 
 }
