+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/ManufactureRepository.java
@@ -3,6 +3,6 @@ package cl.apicom.spring.backend.repository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import cl.apicom.spring.backend.entities.Manufacture;
 
-public interface ManufactureRepository extends PagingAndSortingRepository<Manufacture, Integer>{//Para 
+public interface ManufactureRepository extends PagingAndSortingRepository<Manufacture, Long>{//Para 
 		
 }
