+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/Price_Client_SectorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import cl.apicom.spring.backend.entities.Price_Client_Sector;
 
-public interface Price_Client_SectorRepository extends PagingAndSortingRepository<Price_Client_Sector, Integer>{
+public interface Price_Client_SectorRepository extends PagingAndSortingRepository<Price_Client_Sector, Long>{
 
 }
