+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/Price_Manuf_SectorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import cl.apicom.spring.backend.entities.Price_Manuf_Sector;
 
-public interface Price_Manuf_SectorRepository extends PagingAndSortingRepository<Price_Manuf_Sector, Integer>{
+public interface Price_Manuf_SectorRepository extends PagingAndSortingRepository<Price_Manuf_Sector, Long>{
 
 }
