+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/repository/SectorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import cl.apicom.spring.backend.entities.Sector;
 
-public interface SectorRepository extends PagingAndSortingRepository<Sector, Integer>{
+public interface SectorRepository extends PagingAndSortingRepository<Sector, Long>{
 
 }
